+++ scout/server/blueprints/variants/views.py
@@ -323,18 +323,33 @@ def variant_update(institute_id, case_name, variant_id):
         store.update_acmg(institute_obj, case_obj, user_obj, link, variant_obj, new_acmg)
         flash("updated ACMG classification: {}".format(new_acmg), 'info')
 
+    new_dismiss = request.form.getlist('dismiss_variant')
     if request.form.getlist('dismiss_variant'):
-        new_dismiss = request.form.getlist('dismiss_variant')
         store.update_dismiss_variant(institute_obj, case_obj, user_obj, link, variant_obj,
                                      new_dismiss)
         if new_dismiss:
             flash("Dismissed variant: {}".format(new_dismiss), 'info')
 
-    if variant_obj.get('dismiss_variant') and not request.form.getlist('dismiss_variant'):
-        new_dismiss = request.form.getlist('dismiss_variant')
-        store.update_dismiss_variant(institute_obj, case_obj, user_obj, link, variant_obj,
+    if variant_obj.get('dismiss_variant') and not new_dismiss:
+        if 'dismiss' in request.form:
+            store.update_dismiss_variant(institute_obj, case_obj, user_obj, link, variant_obj,
                                      new_dismiss)
-        flash("Reset variant dismissal: {}".format(variant_obj.get('dismiss_variant')), 'info')
+            flash("Reset variant dismissal: {}".format(variant_obj.get('dismiss_variant')), 'info')
+        else:
+            log.debug("DO NOT reset variant dismissal: {}".format(variant_obj.get('dismiss_variant')), 'info')
+
+    mosaic_tags = request.form.getlist('mosaic_tags')
+    if mosaic_tags:
+        store.update_mosaic_tags(institute_obj, case_obj, user_obj, link, variant_obj,
+                                     mosaic_tags)
+        if new_dismiss:
+            flash("Added mosaic tags: {}".format(mosaic_tags), 'info')
+
+    if variant_obj.get('mosaic_tags') and not mosaic_tags:
+        if 'mosaic' in request.form:
+            store.update_mosaic_tags(institute_obj, case_obj, user_obj, link, variant_obj,
+                                     mosaic_tags)
+            flash("Reset mosaic tags: {}".format(variant_obj.get('mosaic_tags')), 'info')
 
     return redirect(request.referrer)